fix kmeans clustering

+++ android/src/main/java/com/faiss/FaissManager.java
@@ -39,7 +39,7 @@ public class FaissManager {
     return  result;
   }
 
-  public  float[] faissKmeans(String dataFile,int k, int dim, int size) {
+  public  int[] faissKmeans(String dataFile,int k, int dim, int size) {
     return FaissManager.KmeansCluster(dataFile,k,dim,size);
   }
 
@@ -56,5 +56,5 @@ public class FaissManager {
 
   public  static  native  String CreateIndexFromTemplate(int[] idsJ, float[][] vectorsJ, String indexPathJ, byte[] templateIndexJ);
 
-  public  static  native  float[] KmeansCluster(String indexName,int k,int dim,int size);
+  public  static  native  int[] KmeansCluster(String indexName,int k,int dim,int size);
 }
